Mudança de regras para periodo calculado, tira documentos cancelados
